pyproject.toml
[project]
name = "django-comais"
version = "0.1.0"
description = "Sistema de gestão de projetos e documentos do COMAIS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.5.8",
    "django>=5.2.2",
    "django-cors-headers>=4.7.0",
    "django-tinymce>=4.1.0",
    "djangorestframework>=3.16.0",
    "docxtpl>=0.20.0",
    "html2markdown>=0.1.7",
    "ipdb>=0.13.13",
    "markdown>=3.8",
    "openai>=1.84.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
gitadd = "git add ."
gitcommit = "git commit -m 'update'"
gitpush = "git push origin main"
deploy = "ssh comais@192.168.105.13 bash -c '\"cd /app/django_comais &&   git pull && sudo /usr/bin/systemctl restart apache2 \"'"
pushd = "task gitadd && task gitcommit && task gitpush && task deploy"
psh = "task gitadd && task gitcommit && task gitpush"
